renamed to StatusIndicator #184
