+++ frontend/src/App.vue
@@ -23,8 +23,8 @@ import FooterVue from "./components/Footer.vue";
 import { useTextContentStore } from "./store/textContentStore";
 import { provideUiFeedbackStore } from "./store/uiFeedbackStore";
 import UiFeedbackDisplay from "./components/UiFeedbackDisplay.vue";
-import MaintenanceBanner from "./components/MaintenanceBanner.vue";
 import { useConfigStore } from "./store/configStore";
+import StatusIndicator from "./components/StatusIndicator.vue";
 
 const route = useRoute();
 const configStore = useConfigStore();
@@ -42,7 +42,7 @@ onMounted(async () => {
     <AppBarForRegister v-if="route.fullPath == '/register'" />
     <AppBar v-else />
     <v-main>
-      <MaintenanceBanner />
+      <StatusIndicator />
       <RouterView />
     </v-main>
     <FooterVue />
