+++ frontend/src/components/StatusIndicator.vue
